Исправил тест

+++ src/ReportSystem.Test/BaseProfileFixture.cs
@@ -52,7 +52,8 @@ namespace ReportSystem.Test
 		[TearDown]
 		public void Stop()
 		{
-			File.Delete(_fileName);
+			if (File.Exists(_fileName))
+				File.Delete(_fileName);
 			ProfileHelper.End();
 		}
